Shared more data among processed (such as result of the game & symbols)

+++ src/TrisServer.c
@@ -18,32 +18,45 @@ void disposeMemory();
 void disposeSemaphores();
 void initSignals();
 void notifyOpponentReady();
-void notifyPlayerWhoWon(int playerWhoWon);
+void notifyPlayerWhoWonForQuit(int playerWhoWon);
+void notifyGameEnded();
 void notifyServerQuit();
 void exitHandler(int sig);
 void playerQuitHandler(int sig);
 void waitForMove();
 
+// Shared memory
+int *matrix;
 int matId;
-char *matrix;
+
+pid_t *pid;
 int pidId;
-int *pid;
+
+int *result;
+int resId;
+
+char *symbols;
+int symId;
+
+// Semaphores
 int semId;
 
+// Settings
 int timeout;
 char playerOneSymbol;
 char playerTwoSymbol;
 
+// State variables
 bool firstCTRLCPressed = false;
 bool started = false;
+int turn = INITIAL_TURN;
 int playersCount = 0;
-int turn = 1; // Player 1 starts
 
 int main(int argc, char *argv[])
 {
-    if (argc != N_ARGS_SERVER)
+    if (argc != N_ARGS_SERVER + 1)
     {
-        errExit(USAGE_ERROR);
+        errExit(USAGE_ERROR_SERVER);
     }
 
     timeout = atoi(argv[1]);
@@ -58,20 +71,33 @@ int main(int argc, char *argv[])
 
     printf("Inizia il giocatore %d (con PID = %d).\n", turn, pid[turn]);
 
-    while (1)
+    while ((*result = isGameEnded(matrix)) == NOT_FINISHED)
     {
-#ifdef DEBUG
-        printBoard(matrix);
+#if DEBUG
+        printBoard(matrix, playerOneSymbol, playerTwoSymbol);
 #endif
         waitForMove();
     }
 
-    return EXIT_SUCCESS;
+    switch (*result)
+    {
+    case 0:
+        printf(KYEL "\nPareggio.\n");
+        break;
+    case 1:
+    case 2:
+        printf(KGRN "\nVince il giocatore %d (con PID = %d).\n", *result, pid[*result]);
+        break;
+    }
+
+    notifyGameEnded();
+
+    exit(EXIT_SUCCESS);
 }
 
 void init()
 {
-    // Messages
+    // Startup messages
     printWelcomeMessageServer();
     printLoadingMessage();
 
@@ -81,25 +107,36 @@ void init()
     initSharedMemory();
     initSignals();
 
+    // Loading complete
     printLoadingCompleteMessage();
 }
 
 void initSharedMemory()
 {
     matId = getAndInitSharedMemory(MATRIX_SIZE, MATRIX_ID);
-    matrix = (char *)attachSharedMemory(matId);
+    matrix = (int *)attachSharedMemory(matId);
     initBoard(matrix);
 
     pidId = getAndInitSharedMemory(PID_SIZE, PID_ID);
     pid = (int *)attachSharedMemory(pidId);
     initPids(pid);
     setPidAt(pid, 0, getpid());
+
+    resId = getAndInitSharedMemory(RESULT_SIZE, RESULT_ID);
+    result = (int *)attachSharedMemory(resId);
+
+    symId = getAndInitSharedMemory(SYMBOLS_SIZE, SYMBOLS_ID);
+    symbols = (char *)attachSharedMemory(symId);
+    symbols[0] = playerOneSymbol;
+    symbols[1] = playerTwoSymbol;
 }
 
 void disposeMemory()
 {
     disposeSharedMemory(matId);
     disposeSharedMemory(pidId);
+    disposeSharedMemory(resId);
+    disposeSharedMemory(symId);
 }
 
 void initSemaphores()
@@ -108,10 +145,10 @@ void initSemaphores()
 
     short unsigned values[N_SEM];
     values[WAIT_FOR_PLAYERS] = 0;
-    values[WAIT_FOR_OPPONENT] = 0;
+    values[WAIT_FOR_OPPONENT_READY] = 0;
     values[LOCK] = 1;
-    values[PLAYER_ONE_TURN] = 1;
-    values[PLAYER_TWO_TURN] = 0;
+    values[PLAYER_ONE_TURN] = INITIAL_TURN == PLAYER_ONE ? 1 : 0;
+    values[PLAYER_TWO_TURN] = INITIAL_TURN == PLAYER_TWO ? 1 : 0;
     values[WAIT_FOR_MOVE] = 0;
 
     setSemaphores(semId, N_SEM, values);
@@ -179,21 +216,27 @@ void playerQuitHandler(int sig)
     int playerWhoQuitted = sig == SIGUSR1 ? PLAYER_ONE : PLAYER_TWO;
     int playerWhoStayed = playerWhoQuitted == PLAYER_ONE ? PLAYER_TWO : PLAYER_ONE;
 
-    printf("\nIl giocatore %d (con PID = %d) ha abbandonato la partita.\n", playerWhoQuitted, pid[playerWhoQuitted]);
+    printf("\nIl giocatore %d (con PID = %d) ha abbandonato la partita.", playerWhoQuitted, pid[playerWhoQuitted]);
     setPidAt(pid, playerWhoQuitted, 0);
     playersCount--;
 
     if (started)
     {
-        printf("Partita terminata: vince il giocatore %d (con PID = %d).\n", playerWhoStayed, pid[playerWhoStayed]);
-        notifyPlayerWhoWon(playerWhoStayed);
+        *result = QUIT;
+        printf(" Vince il giocatore %d (con PID = %d).\n", playerWhoStayed, pid[playerWhoStayed]);
+        notifyPlayerWhoWonForQuit(playerWhoStayed);
         exit(EXIT_SUCCESS);
+    }
+    else
+    {
+        printf("\n");
     }
 }
 
 void waitForPlayers()
 {
     printf("\nAttendo i giocatori...\n");
+
     while (playersCount < 2)
     {
         do
@@ -211,24 +254,40 @@ void waitForPlayers()
 
 void notifyOpponentReady()
 {
-    signalSemaphore(semId, WAIT_FOR_OPPONENT, 2);
+    signalSemaphore(semId, WAIT_FOR_OPPONENT_READY, 2);
 }
 
-void notifyPlayerWhoWon(int playerWhoWon)
+void notifyPlayerWhoWonForQuit(int playerWhoWon)
 {
     kill(pid[playerWhoWon], SIGUSR2);
 }
 
-void notifyServerQuit()
+void notifyGameEnded()
 {
-    kill(pid[PLAYER_ONE], SIGUSR1);
-    kill(pid[PLAYER_TWO], SIGUSR1);
+    kill(pid[PLAYER_ONE], SIGUSR2);
+    kill(pid[PLAYER_TWO], SIGUSR2);
 }
 
-void waitForMove(){
+void notifyServerQuit()
+{
+    if (pid[PLAYER_ONE] != 0)
+        kill(pid[PLAYER_ONE], SIGUSR1);
+
+    if (pid[PLAYER_TWO] != 0)
+        kill(pid[PLAYER_TWO], SIGUSR1);
+}
+
+void waitForMove()
+{
     printf("In attesa della mossa del giocatore %d (con PID = %d)...\n", turn, pid[turn]);
-    waitSemaphore(semId, WAIT_FOR_MOVE, 1);
+
+    do
+    {
+        errno = 0;
+        waitSemaphore(semId, WAIT_FOR_MOVE, 1);
+    } while (errno == EINTR);
     printf("Mossa ricevuta dal giocatore %d (con PID = %d).\n", turn, pid[turn]);
+
     turn = turn == 1 ? 2 : 1;
     signalSemaphore(semId, turn, 1);
 }
